% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwC.R
\name{emof}
\alias{emof}
\title{Make the Darwin Core extended MeasurementOrFact table}
\usage{
emof(df)
}
\arguments{
\item{df}{dataframe with unique IDs for events}
}
\value{
darwin core compatible eMoF table
}
\description{
Make the Darwin Core extended MeasurementOrFact table
}
