% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwC.R
\name{event}
\alias{event}
\title{Make the Darwin Core Event table}
\usage{
event(df)
}
\arguments{
\item{df}{dataframe with unique IDs for events}
}
\value{
darwin core compatible event table
}
\description{
Make the Darwin Core Event table
}
