% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{gather_measures}
\alias{gather_measures}
\title{Gathers the fish length/abundance measurements to a long tidy r dataframe}
\usage{
gather_measures(df)
}
\arguments{
\item{df}{RLS dataframe imported using the readRLS function}
}
\value{
r dataframe
}
\description{
Gathers the fish length/abundance measurements to a long tidy r dataframe
}
