% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwC.R
\name{makeIDs}
\alias{makeIDs}
\title{Add unique ids for events and species occurrence}
\usage{
makeIDs(df)
}
\arguments{
\item{df}{long dataframe with species names verified}
}
\value{
dataframe with unqiue ids for events and species occurrences
}
\description{
Add unique ids for events and species occurrence
}
