% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DwC.R
\name{occurence}
\alias{occurence}
\title{Make the Darwin Core occurence table}
\usage{
occurence(df)
}
\arguments{
\item{df}{dataframe with unique IDs for occurrences}
}
\value{
darwin core compatible occurence table
}
\description{
Make the Darwin Core occurence table
}
