% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{readRLS}
\alias{readRLS}
\title{Load a RLS workbook to a R data frame}
\usage{
readRLS(xlsx, sheetname = "Data")
}
\arguments{
\item{xlsx}{path to the reef life survey excel workbook}

\item{sheetname}{name of the sheet (tab) with the data}
}
\value{
r dataframe
}
\description{
Imports the data tab for a reef life survey workbook into a r dataframe
}
