% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{renameRLS}
\alias{renameRLS}
\title{Internal function to rename columns to match Darwin Core terms}
\usage{
renameRLS(df)
}
\arguments{
\item{df}{RLS dataframe imported using readRLS function}
}
\value{
dataframe with column names renamed
}
\description{
Internal function to rename columns to match Darwin Core terms
}
