% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{replace_sciName}
\alias{replace_sciName}
\title{Replaces all occurnaces of a scientific name with a new value}
\usage{
replace_sciName(df, old, new)
}
\arguments{
\item{df}{dataframe with species names to validate (col should be named scientificname)}

\item{old}{scientific name to replace}

\item{new}{the new scientific name to use}
}
\value{
dataframe
}
\description{
Intended use is to make your data complient with WoRMS
}
