% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{uniqueSites}
\alias{uniqueSites}
\title{Parse unique site names from RLS datasheet}
\usage{
uniqueSites(df)
}
\arguments{
\item{df}{RLS dataframe}
}
\value{
r dataframe with unqiue site names, site ids, latitudes, longitudes
}
\description{
Parse unique site names from RLS datasheet
}
