% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_species}
\alias{verify_species}
\title{Verify species names using the World Register of Marine Species (WoRMS)}
\usage{
verify_species(df, allowUnmatched = FALSE)
}
\arguments{
\item{df}{dataframe with species names to validate (col should be named scientificname)}

\item{allowUnmatched}{boolean to allow unmatched records (default is false)}
}
\value{
dataframe with matches joined to the rows
}
\description{
Verify species names using the World Register of Marine Species (WoRMS)
}
